+++ yash-cli/CHANGELOG.md
@@ -9,6 +9,13 @@ used by other programs.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.3] - Unreleased
+
+### Changed
+
+- The `times` built-in now uses the `getrusage` system call instead of `times`
+  to get CPU times. This provides better resolution on many systems.
+
 ## [3.0.2] - 2025-10-16
 
 ### Fixed
@@ -286,6 +293,7 @@ later.
 
 - Initial release of the shell
 
+[3.0.3]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.3
 [3.0.2]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.2
 [3.0.1]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.1
 [3.0.0]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.0